fix: shells computations refactor: Deleting shells.py for typing
