+++ scripts/block3d.py
@@ -20,15 +20,20 @@ box_red = primitives3d.Block(
                vm.Vector3D(0, 0.4, 0), vm.Vector3D(0, 0, 0.4)),
     color=(0.2, 1, 0.4), alpha=0.6)
 
-ray = vm.edges.LineSegment3D(vm.Point3D(-0.15, -0.15, -0.15),
-                           vm.Point3D(0.34416933003412165,
-                                      0.3660556568489298,
-                                      0.6104289251777456))
+p1_ray = vm.Point3D(-0.15, -0.15, -0.15)
+p2_ray = vm.Point3D(0.009855980224206917, 0.6250574317556334, -0.1407142090413507)
+ray = vm.edges.LineSegment3D(p1_ray, p2_ray)
 
-ax = ray.plot()
+
+ax = ray.plot(color='b')
+p1_ray.plot(ax=ax, color='b')
+p2_ray.plot(ax=ax, color='b')
 box_red.plot(ax=ax, color='r')
-for inter_point in box_red.linesegment_intersections(ray):
-    inter_point.plot(ax=ax, color='r')
+for face, inter_points in box_red.linesegment_intersections(ray):
+    # print('ip', inter_point)
+    face.plot(ax=ax, color='b')
+    for inter_point in inter_points:
+        inter_point.plot(ax=ax, color='r')
 
 
 box_red.color = (1, 0.1, 0.1)
@@ -55,12 +60,12 @@ model = vm.core.VolumeModel([box, box_red, box_green, box_blue])
 model.babylonjs(debug=True)
 print('\n\n@@@@@@@\n')
 assert box.is_inside_shell(box_red, resolution) == True
-# assert box_red.is_inside_shell(box, resolution) == False
-#
-# assert box.is_inside_shell(box_green, resolution) == False
-# assert box_green.is_inside_shell(box, resolution) == False
-#
-# assert box.is_inside_shell(box_blue, resolution) == False
-# assert box_blue.is_inside_shell(box, resolution) == False
+assert box_red.is_inside_shell(box, resolution) == False
+
+assert box.is_inside_shell(box_green, resolution) == False
+assert box_green.is_inside_shell(box, resolution) == False
+
+assert box.is_inside_shell(box_blue, resolution) == False
+assert box_blue.is_inside_shell(box, resolution) == False
